fix AudioNode delayed disable race condition
When AudioNode.disableOutputsIfNecessary is called, it calls
this.getTailTime() which returns a "delay" time before the node is
actually disabled.

If the the node is enabled again within that time, we need to clear the
disable countdown (this is the bug fix).

Also rename "suicideAtSampleFrame" to "disableAtSampleFrame" because it's
a crude name and it isn't even an accurate description.